Move offset reset to SL_Init, which runs every game cycle.
Also store the default offset in ThemePrefs. Currently nothing is
allowed to modify this, so the offset will return to that value before
every credit.

NOTE: With the current implementation of gnGlobalOffset, values outside
of -0.15;0.15 will be reset to 0 when SL-PlayerOptions runs. This will
be fixed when the options become relative to the current offset.